fix: change ClvSerializeTerm from uint64_t to uint16_t

+++ src/include/conclave-serialize/types.h
@@ -16,7 +16,7 @@ typedef uint64_t ClvSerializeUserSessionId;
 typedef uint8_t ClvSerializeRoomConnectionIndex;
 typedef uint64_t ClvSerializeApplicationId;
 typedef uint64_t ClvSerializeClientNonce;
-typedef uint64_t ClvSerializeTerm;
+typedef uint16_t ClvSerializeTerm;
 typedef uint64_t ClvSerializeKnowledge;
 
 typedef enum ClvSerializeConnectedToOwnerState {
